[proto] make pkg/proto into a mod + fix moved pbutils
