+++ pkg/tagger/server/server.go
@@ -14,11 +14,11 @@ import (
 	"google.golang.org/grpc/status"
 
 	pb "github.com/DataDog/datadog-agent/pkg/proto/pbgo"
-	pbutils "github.com/DataDog/datadog-agent/pkg/proto/utils"
 	"github.com/DataDog/datadog-agent/pkg/tagger"
 	"github.com/DataDog/datadog-agent/pkg/tagger/telemetry"
 	"github.com/DataDog/datadog-agent/pkg/util/grpc"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
+	pbutils "github.com/DataDog/datadog-agent/pkg/util/proto"
 )
 
 const (
