+++ pkg/util/proto/workloadmeta_test.go
@@ -3,7 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-2020 Datadog, Inc.
 
-package utils
+package proto
 
 import (
 	"testing"
